Converted vpgl_convert_local_to_global_coordinates from float to double

+++ contrib/brl/bpro/core/vpgl_pro/processes/vpgl_convert_local_to_global_coordinates_process.cxx
@@ -13,26 +13,26 @@ bool vpgl_convert_local_to_global_coordinates_process_cons(bprb_func_process& pr
 {
   //this process takes four inputs: 
   // 0: (string) lvcs filename
-  // 1: (float)  latitude
-  // 2: (float)  longitude
-  // 3: (float)  elevation
+  // 1: (double)  latitude
+  // 2: (double)  longitude
+  // 3: (double)  elevation
   bool ok=false;
   vcl_vector<vcl_string> input_types;
   input_types.push_back("vpgl_lvcs_sptr"); 
-  input_types.push_back("float");
-  input_types.push_back("float");
-  input_types.push_back("float");
+  input_types.push_back("double");
+  input_types.push_back("double");
+  input_types.push_back("double");
   ok = pro.set_input_types(input_types);
   if (!ok) return ok;
 
   // outputs x,y,z in local coordinates
-  // 0: (float) x
-  // 1: (float) y
-  // 2: (float) z
+  // 0: (double) x
+  // 1: (double) y
+  // 2: (double) z
   vcl_vector<vcl_string> output_types;
-  output_types.push_back("float");  
-  output_types.push_back("float");  
-  output_types.push_back("float");  
+  output_types.push_back("double");  
+  output_types.push_back("double");  
+  output_types.push_back("double");  
   ok = pro.set_output_types(output_types);
   if (!ok) return ok;
   
@@ -50,17 +50,17 @@ bool vpgl_convert_local_to_global_coordinates_process(bprb_func_process& pro)
 
   // get the inputs
   vpgl_lvcs_sptr lvcs = pro.get_input<vpgl_lvcs_sptr>(0);
-  float x = pro.get_input<float>(1);
-  float y = pro.get_input<float>(2);
-  float z = pro.get_input<float>(3);
+  double x = pro.get_input<double>(1);
+  double y = pro.get_input<double>(2);
+  double z = pro.get_input<double>(3);
 
  
   double lat,lon,el;
   lvcs->local_to_global(x,y,z,vpgl_lvcs::wgs84,lon,lat,el,vpgl_lvcs::DEG, vpgl_lvcs::METERS);
 
-  pro.set_output_val<float>(0, (float)lat);
-  pro.set_output_val<float>(1, (float)lon);
-  pro.set_output_val<float>(2, (float)el);
+  pro.set_output_val<double>(0, (double)lat);
+  pro.set_output_val<double>(1, (double)lon);
+  pro.set_output_val<double>(2, (double)el);
 
   return true;
 }
